initial commit
The current state is a stripped down version of an old project to be
used as a starting point. It should compile but will not link as
there are no definitions for the main library, only the old public interface.

I have included an example project (n-puzzle) that compiled and ran successfully
with all the current configuration before I stripped the project down.
Eventually it should compile again, once the library is fleshed out, and
perhaps with a bit of modification.